# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Danjjak API
  version: 0.1.0
  description: 고령 사용자의 금융 업무 흐름을 지원하는 단짝 서비스 API입니다.
servers:
  - url: http://localhost:8080/danjjak
    description: 로컬 Tomcat
security: []
...
paths:
...
  /api/support/guardian:
    put:
      tags: [Support]
      operationId: updateGuardianContact
      summary: 사용자당 보호자 연락처 1개 수정
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuardianContactRequest'
      responses:
        '200':
          description: 수정된 보호자 연락처
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuardianContact'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/anomaly-events/{anomalyEventId}/guardian-notification:
    post:
      tags: [Support]
      operationId: sendGuardianNotification
      summary: 사용자가 선택한 HIGH 이상거래 카카오 알림 전송
      description: 보호자 전화번호가 아니라 로그인한 시연 계정의 카카오 나에게 보내기를 사용한다.
      parameters:
        - $ref: '#/components/parameters/AnomalyEventId'
      responses:
        '200':
          description: 실제 발송 또는 Mock 대체 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
components:
  parameters:
...
    AnomalyEventId:
      name: anomalyEventId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: 현재 데이터 상태와 충돌
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: RESOURCE_NOT_FOUND
        message:
          type: string
          example: 요청한 리소스를 찾을 수 없습니다.
...
    GuardianContact:
      type: object
      required: [guardianContactId, phoneNumber]
      properties:
        guardianContactId:
          type: integer
          format: int64
        phoneNumber:
          type: string
          pattern: '^[0-9-]{8,20}$'
    GuardianContactRequest:
      type: object
      required: [phoneNumber]
      properties:
        phoneNumber:
          type: string
          pattern: '^[0-9-]{8,20}$'
...
    NotificationResponse:
      type: object
      required:
        [anomalyEventId, deliveryMode, result, actualAttempted, actualSucceeded, detail]
      properties:
        anomalyEventId:
          type: integer
          format: int64
        deliveryMode:
          type: string
          enum: [ACTUAL, MOCK]
        result:
          type: string
          enum: [SENT, MOCKED_NO_TOKEN, MOCKED_AFTER_ACTUAL_FAILURE]
        actualAttempted:
          type: boolean
        actualSucceeded:
          type: boolean
        detail:
          type: string